Added icon images
These are used by the PRISMSDK 5.2 make file when compiling to give the program an image icon when opened on the calculator. I aimed to design them in a similar style to the programs that come with the calculator to make them fit with the CASIO style. They are 8 bit RGB (24 bit total) bitmaps with dimensions 92 x 64 if anyone wants to customise them.